% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GephiReadyFile.R
\name{build_gephi_adjacency}
\alias{build_gephi_adjacency}
\title{Create Gephi Ready File for Twitter Social Media Network Analysis}
\usage{
build_gephi_adjacency(
  query,
  numOfTweets,
  twitter_api_key,
  twitter_secret,
  app_name
)
}
\arguments{
\item{query}{character string}

\item{numOfTweets}{integer}

\item{twitter_api_key}{character string}

\item{twitter_secret}{character string}

\item{app_name}{character string}
}
\value{
dataframe
}
\description{
Create Gephi Ready File for Twitter Social Media Network Analysis
}
